+++ lightsim2grid/tests/test_GridModel.py
@@ -11,7 +11,7 @@ import numpy as np
 import pandapower.networks as pn
 import pandapower as pp
 
-from lightsim2grid.initGridModel import init
+from lightsim2grid.gridmodel import init
 import warnings
 import pdb